requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2
django-admin-interface==0.30.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
html5lib==1.1
idna==3.10
inflection==0.5.1
lxml==5.4.0
numpy==2.4.0
oscrypto==1.3.0
packaging==25.0
pandas==2.2.2
pillow==11.2.1
pycparser==2.22
pyHanko==0.27.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
razorpay==2.0.0
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
xhtml2pdf==0.2.17
